[CHANGE] lazy load hcaptcha and move it to a separate file
